Fix documentation comments
