+++ src/business_logic/transaction/objects/interal_declare.rs
@@ -127,6 +127,8 @@ impl InternalDeclare {
         Ok(())
     }
 
+    /// Execute a call to the cairo-vm using the accounts_validation.cairo contract to validate
+    /// the contract that is being declared. Then it returns the transaction execution info of the run.
     pub fn apply<S: Default + State + StateReader + Clone>(
         &self,
         state: &mut S,
@@ -262,6 +264,8 @@ impl InternalDeclare {
         Ok(())
     }
 
+    /// Calculates actual fee used by the transaction using the execution
+    /// info returned by apply(), then updates the transaction execution info with the data of the fee.  
     pub fn execute<S: Default + State + StateReader + Clone>(
         &self,
         state: &mut S,
