+++ src/business_logic/transaction/objects/internal_declare.rs
@@ -127,6 +127,8 @@ impl InternalDeclare {
         Ok(())
     }
 
+    /// Executes a call to the cairo-vm using the accounts_validation.cairo contract to validate 
+    /// the contract that is being declared. Then it returns the transaction execution info of the run. 
     pub fn apply<S: Default + State + StateReader + Clone>(
         &self,
         state: &mut S,
@@ -163,7 +165,6 @@ impl InternalDeclare {
     // ~~~~~~~~~~~~~~~~~~~~~~~~~~~
     // Internal Account Functions
     // ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
     pub fn get_execution_context(&self, n_steps: u64) -> TransactionExecutionContext {
         TransactionExecutionContext::new(
             self.sender_address.clone(),
@@ -211,10 +212,8 @@ impl InternalDeclare {
         Ok(Some(call_info))
     }
 
-    // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-    // Calculates and charges the actual fee.
-    // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+    /// Calculates and charges the actual fee.
     pub fn charge_fee<S: Default + State + StateReader + Clone>(
         &self,
         state: &mut S,
@@ -261,7 +260,8 @@ impl InternalDeclare {
 
         Ok(())
     }
-
+    /// Calculates actual fee used by the transaction using the execution 
+    /// info returned by apply(), then updates the transaction execution info with the data of the fee.  
     pub fn execute<S: Default + State + StateReader + Clone>(
         &self,
         state: &mut S,
